Fixed response type

+++ entities.go
@@ -121,7 +121,7 @@ type MatchDTO struct {
 				EpicMonsterKillsWithin30SecondsOfSpawn   int     `json:"epicMonsterKillsWithin30SecondsOfSpawn"`
 				EpicMonsterSteals                        int     `json:"epicMonsterSteals"`
 				EpicMonsterStolenWithoutSmite            int     `json:"epicMonsterStolenWithoutSmite"`
-				FirstTurretKilled                        int     `json:"firstTurretKilled"`
+				FirstTurretKilled                        float64 `json:"firstTurretKilled"`
 				FlawlessAces                             int     `json:"flawlessAces"`
 				FullTeamTakedown                         int     `json:"fullTeamTakedown"`
 				GameLength                               float64 `json:"gameLength"`
